Updates for nf-core tools v2
The 'samplesheet' parameter was renamed to 'input' for consistency with nf-core pipelines.

The celltype markers generated during the finalize step now saves cellmarker plots and tables.  An additional parameter, top_n, was exposed to NF and allows the number of genes used for the plots/tables to be specified (default: 5).

Workflow, parameter, and parameter schema changes were made for the above changes.