Fixed celltype marker file output locations and DGE errorStrategy
Celltype and cluster markers are now correctly published by the pipeline with file names better reflecting the cell groupings used.

The DGE errorStrategy was updated to 'ignore'.  This allows for the possibility that a valid model may fail to fit and trigger and error on a cell-type not represented among different classes of the dependent variable.