Added software version scraping process for R packages
The nf-core lint tool requires a software version scraping process.  A new NF process now runs a R script (inside the container) which retrieves and saves all installed package versions as a .tsv file which is exported to the pipeline_info folder.  The versions are prefixed with v (e.g. v2.0.1) for consistency with nf-core.